+++ tg_bot/docker/.dockerignore
